+++ src/safeds/data/tabular/containers/__init__.py
@@ -6,10 +6,10 @@ import apipkg
 
 if TYPE_CHECKING:
     from ._column import Column
-    from ._experimental_polars_cell import ExperimentalPolarsCell
-    from ._experimental_polars_column import ExperimentalPolarsColumn
-    from ._experimental_polars_row import ExperimentalPolarsRow
-    from ._experimental_polars_table import ExperimentalPolarsTable
+    from ._experimental_cell import ExperimentalPolarsCell
+    from ._experimental_column import ExperimentalPolarsColumn
+    from ._experimental_row import ExperimentalPolarsRow
+    from ._experimental_table import ExperimentalPolarsTable
     from ._row import Row
     from ._table import Table
 
@@ -17,10 +17,10 @@ apipkg.initpkg(
     __name__,
     {
         "Column": "._column:Column",
-        "ExperimentalPolarsCell": "._experimental_polars_cell:ExperimentalPolarsCell",
-        "ExperimentalPolarsColumn": "._experimental_polars_column:ExperimentalPolarsColumn",
-        "ExperimentalPolarsRow": "._experimental_polars_row:ExperimentalPolarsRow",
-        "ExperimentalPolarsTable": "._experimental_polars_table:ExperimentalPolarsTable",
+        "ExperimentalPolarsCell": "._experimental_cell:ExperimentalPolarsCell",
+        "ExperimentalPolarsColumn": "._experimental_column:ExperimentalPolarsColumn",
+        "ExperimentalPolarsRow": "._experimental_row:ExperimentalPolarsRow",
+        "ExperimentalPolarsTable": "._experimental_table:ExperimentalPolarsTable",
         "Row": "._row:Row",
         "Table": "._table:Table",
     },
