+++ src/safeds/data/tabular/containers/_experimental_cell.py
