+++ src/safeds/data/tabular/containers/_experimental_row.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Any
 if TYPE_CHECKING:
     from safeds.data.tabular.typing import ColumnType, Schema
 
-    from ._experimental_polars_cell import ExperimentalPolarsCell
+    from ._experimental_cell import ExperimentalPolarsCell
 
 
 class ExperimentalPolarsRow(ABC, Mapping[str, Any]):
